Mark flaky logging test as xfail
