+++ mypy.ini
@@ -49,6 +49,9 @@ ignore_errors = True
 [mypy-ume.snapshot_routes]
 ignore_errors = True
 
+[mypy-ume.ledger_routes]
+ignore_errors = True
+
 [mypy-ume.config]
 ignore_errors = True
