+++ src/ume/dashboard_routes.py
@@ -7,7 +7,7 @@ from fastapi import APIRouter, Depends
 from .audit import get_audit_entries
 from .graph_adapter import IGraphAdapter
 from .api_deps import get_current_role, get_graph, get_vector_store
-from .vector_store import VectorStore
+from .vector_backends import VectorStore
 
 router = APIRouter(prefix="/dashboard")
