+++ src/ume/vector_backends/__init__.py
@@ -1,0 +1,12 @@
+"""Vector store backend implementations."""
+
+from .base import VectorStore, VectorBackend
+from .faiss_backend import FaissBackend
+from .chroma_backend import ChromaBackend
+
+__all__ = [
+    "VectorStore",
+    "VectorBackend",
+    "FaissBackend",
+    "ChromaBackend",
+]
